refactor(validation): modularize ValidatorRegistry.js (1,234 → 10 modules)
SD-LEO-REFACTOR-VALIDATOR-REG-001 - Child 9 of orchestrator SD-LEO-REFACTOR-LARGE-FILES-002

Extracted 1,234 LOC validator registry into focused modules by gate:

Core (1 file):
- core.js - ValidatorRegistry class with core methods (register, get, has, etc.)

Gates (8 files):
- gate-l-sd-creation.js - 6 validators (SD existence, objectives, priority, etc.)
- gate-1-plan-to-exec.js - 10 validators (PRD quality, user stories, sub-agents)
- gate-2-implementation-fidelity.js - 11 validators (UI, migrations, RLS, E2E)
- gate-3-traceability.js - 5 validators (recommendation adherence, quality)
- gate-4-strategic-value.js - 4 validators (value delivered, patterns)
- gate-q-quality.js - 5 validators (executive summary, decisions, issues)
- additional-validators.js - 14 validators (transition, branch, retrospective)
- index.js - Gate exports

Main (1 file):
- index.js - Creates registry and registers all validators

Original file now thin re-export (21 LOC). 52 validators registered, 11 exports.
