+++ scripts/modules/handoff/validation/validator-registry/gates/index.js
@@ -1,0 +1,12 @@
+/**
+ * Gate Validators Index
+ * Part of SD-LEO-REFACTOR-VALIDATOR-REG-001
+ */
+
+export { registerGateLValidators } from './gate-l-sd-creation.js';
+export { registerGate1Validators } from './gate-1-plan-to-exec.js';
+export { registerGate2Validators } from './gate-2-implementation-fidelity.js';
+export { registerGate3Validators } from './gate-3-traceability.js';
+export { registerGate4Validators } from './gate-4-strategic-value.js';
+export { registerGateQValidators } from './gate-q-quality.js';
+export { registerAdditionalValidators } from './additional-validators.js';
